refactor(mainpage): Organize API functions and type imports (#12)
* refactor: Replace category codes with constants

* chore: Move and extend API types

* refactor: Moved API functions and update type imports
