+++ constants/categoryCode.ts
@@ -1,0 +1,4 @@
+export const CATEGORY_CODE = {
+  hospital: "HP8",
+  pharmacy: "PM9",
+};
